examples/quickstart/requirements.txt
graphiti-core
python-dotenv>=1.0.0
pyairtable